lambda/requirements.txt
ask-sdk-core
ask-sdk-dynamodb-persistence-adapter
requests